Fix mouse offset error when canvas has padding
This is a error that occurs if you try to use Perlenspiel on a custom
html page with non-zero padding values for the canvas.
What happens is that the mouse detection point is offset by the amount
of the padding along the top and left edges.
This fix calculates an adjusted offset when the engine initializes, so
changes after it loads won't be reflected, but it also won't waste time
recalculating for each bead.